Make initializer API match docs
I just wrote in the README that initializer methods in the GameOfLife
can accept **kwargs, because that makes sense, but it wasn't actually
true.

Previously, keyword args passed to GameOfLife() would be saved to
self.other_args and accessed from there in intializer methods.

Now, these args are passed down to initializer methods, which are
expected to define their relevant optional args as keyword args.