Make another attempt to reduce flickering
This attempts to avoid repeated allocations by creating a buffer up
front. It also localizes all names in the grid-printing function.

This still doesn't remove flickering on iTerm2. According to some brief
research, it seems like this might be caused by the terminal emulator
and not by performance problems in the code or OS buffering. Some
empirical evidence for this is that I'm not seeing any flickering when
using macOS's built-in Terminal app.